Split out requestThemes
